requirements.txt
# Optional: regenerate assets/images/pdf-pages from the PDF
pymupdf>=1.24.0
Pillow>=10.0.0
